object picking

+++ 300/_ENGINE/include/Graphics/GraphicsSystem.h
@@ -36,7 +36,7 @@ public:
 	// -- FBO --
 	unsigned int GetGameAttachment()		{ return m_Fbo.GetGameAttachment(); }
 	unsigned int GetEditorAttachment()		{ return m_Fbo.GetEditorAttachment(); }
-	unsigned int GetEntityID(int x, int y)	{ return m_Fbo.ReadEntityID(x, y); }
+	unsigned int GetEntityID(float x, float y)	{ return m_Fbo.ReadEntityID(x, y); }
 
 	// -- Getter --
 	GFX::DebugRenderer& getDebugRenderer() { return m_Renderer; }
